+++ src/core/materials/BasicMaterial.ts
@@ -37,55 +37,15 @@ export class BasicMaterial extends AbstractMaterial {
     }
   }
 
-  /** @inheritdoc */
   public override getRenderManifest(): RenderManifest {
     if (undefined === this._renderManifest) {
-      this._renderManifest = {
-        shaderId: this.type,
-        properties: {
-          u_color: this.color.toFloat32Array(),
-          u_specColor: new Float32Array([1, 1, 1, 1]),
-          u_texOffset: [0, 0],
-          u_texRepeat: [1, 1],
-          u_shininess: 32.0,
-          u_isTerrain: 0.0,
-          u_metallic: 0.0,
-          u_roughness: 0.5,
-          u_extraParams: [1.0, 0, 0, 0],
-          u_liquidParams: [0, 0, 0, 0],
-          u_thresholds: [0, 0, 0, 0],
-        },
-        textures: {
-          u_diffuseMap: this.diffuseMap,
-        },
-      };
+      this._renderManifest = this._createBaseManifest();
     }
 
-    const props = this._renderManifest.properties as Record<string, unknown>;
-    const texs = this._renderManifest.textures as Record<string, unknown>;
+    this._syncBaseManifestState();
+    this._syncTexOffsetRepeat(this.diffuseMap);
 
-    props["u_color"] = this.color.toFloat32Array();
-    if (this.diffuseMap) {
-      (props["u_texOffset"] as number[])[0] = this.diffuseMap.offset.x;
-      (props["u_texOffset"] as number[])[1] = this.diffuseMap.offset.y;
-      (props["u_texRepeat"] as number[])[0] = this.diffuseMap.repeat.x;
-      (props["u_texRepeat"] as number[])[1] = this.diffuseMap.repeat.y;
-      texs["u_diffuseMap"] = this.diffuseMap;
-    } else {
-      (props["u_texOffset"] as number[])[0] = 0;
-      (props["u_texOffset"] as number[])[1] = 0;
-      (props["u_texRepeat"] as number[])[0] = 1;
-      (props["u_texRepeat"] as number[])[1] = 1;
-      texs["u_diffuseMap"] = undefined;
-    }
-
-    this._renderManifest.state = {
-      ...this._renderManifest.state,
-      culling: this.cullMode,
-      depthWrite: this.depthWrite,
-      depthTest: this.depthTest,
-      transparent: this.transparent,
-    };
+    this._renderManifest.textures["u_diffuseMap"] = this.diffuseMap;
 
     return this._renderManifest;
   }