+++ src/core/materials/LambertMaterial.ts
@@ -56,54 +56,20 @@ export class LambertMaterial extends AbstractMaterial {
   /** @inheritdoc */
   public override getRenderManifest(): RenderManifest {
     if (undefined === this._renderManifest) {
-      this._renderManifest = {
-        shaderId: this.type,
-        properties: {
-          u_color: this.color.toFloat32Array(),
-          u_specColor: new Float32Array([1, 1, 1, 1]),
-          u_texOffset: [0, 0],
-          u_texRepeat: [1, 1],
-          u_shininess: 32.0,
-          u_isTerrain: 0.0,
-          u_metallic: 0.0,
-          u_roughness: 1.0,
-          u_extraParams: [1.0, 0, this.normalScale.x, this.normalScale.y],
-          u_liquidParams: [0, 0, 0, 0],
-          u_thresholds: [0, 0, 0, 0],
-        },
-        textures: {
-          u_diffuseMap: this.diffuseMap,
-          u_normalMap: this.normalMap,
-        },
-      };
+      this._renderManifest = this._createBaseManifest();
     }
+
+    this._syncBaseManifestState();
+    this._syncTexOffsetRepeat(this.diffuseMap);
 
     const props = this._renderManifest.properties as Record<string, unknown>;
     const texs = this._renderManifest.textures as Record<string, unknown>;
 
-    props["u_color"] = this.color.toFloat32Array();
     (props["u_extraParams"] as number[])[2] = this.normalScale.x;
     (props["u_extraParams"] as number[])[3] = this.normalScale.y;
 
-    if (this.diffuseMap) {
-      (props["u_texOffset"] as number[])[0] = this.diffuseMap.offset.x;
-      (props["u_texOffset"] as number[])[1] = this.diffuseMap.offset.y;
-      (props["u_texRepeat"] as number[])[0] = this.diffuseMap.repeat.x;
-      (props["u_texRepeat"] as number[])[1] = this.diffuseMap.repeat.y;
-    } else {
-      (props["u_texOffset"] as number[])[0] = 0;
-      (props["u_texOffset"] as number[])[1] = 0;
-      (props["u_texRepeat"] as number[])[0] = 1;
-      (props["u_texRepeat"] as number[])[1] = 1;
-    }
-
     texs["u_diffuseMap"] = this.diffuseMap;
     texs["u_normalMap"] = this.normalMap;
-
-    this._renderManifest.state = {
-      ...this._renderManifest.state,
-      culling: this.cullMode,
-    };
 
     return this._renderManifest;
   }