+++ src/renderers/passes/index.ts
@@ -1,3 +1,3 @@
-/// src/renderers/passes/index.ts
 export * from "./MainRenderPass.js";
 export * from "./PostProcessPass.js";
+export * from "./CascadedShadowPassGPU.js";